feat: rename files in `content` to match kebab-case
