+++ pages/content/src/sample-function.ts
